Move hosting code to Service.
